Upload 4 questions & codes
I copied the 4 inqueries and codes for the Udacity Sakila Movie 
Databases project 1.  I decided to upload them as a text entry because I 
am unsure how to write html just yet and I want to be able to copy it 
later and read it as I wrote it.  

Inqueries appear on the following lines: #1, #18, #38, #63